refactor: move all 2d math to src/lib/math2d, code cleanup to remove circular dependencies
