+++ src/components/SimpleProfile/index.tsx
@@ -8,20 +8,14 @@
 
 // Points generated by point sets are sorted by angle and will be used for profile curve generation.
 
+import { SizeUnit } from "@/lib/units";
+
+import { PointSet } from "./types";
 import { calculateDrawProps, generateProfile } from "./util";
-import { SizeUnit } from "@/lib/types";
 import { GuideLines } from "./GuideLines";
 import { GuideArc } from "./GuideArc";
 import { ProfilePoints } from "./ProfilePoints";
 import { ProfileCurve } from "./ProfileCurve";
-
-export type PointSet = {
-  offset: number;
-  count: number;
-  angleStart: number;
-  angleStep: number;
-  color: string;
-};
 
 export type SimpleProfileProps = {
   sections: number;
