+++ src/components/SimpleProfile/util.ts
@@ -1,66 +1,54 @@
-import { SizeUnit } from "@/lib/types";
-import { PointSet } from ".";
-import { generateProfileSectionCurve } from "./curve";
+import {
+  Vec2,
+  catmullRomCurvePoint,
+  distanceSqr,
+  estimateCatmullRomCurveLength,
+  pointOnCircle,
+  rotate,
+  simplifyProfilePoints,
+} from "@/lib/math2d";
 
-export type Vec2 = {
-  x: number;
-  y: number;
-};
+import {
+  SizeUnit,
+  minDistanceForUnit,
+  minSimplifyAreaForUnit,
+} from "@/lib/units";
 
-export type ProfilePoint = {
-  position: Vec2;
-  angle: number;
-  priority: number;
-  color: string;
-};
+import { DrawProps, PointSet, Profile, ProfilePoint } from "./types";
 
-export type DrawProps = {
-  strokeWidth: number;
-  viewBox: string;
-  sections: number;
-  angleStart: number;
-  angleStep: number;
-  profileRadius: number;
-  maxOffset: number;
-};
+export const ANGLE_EPSILON = (0.01 * (Math.PI * 2)) / 360; // 0.01 deg
 
-export type Profile = {
-  referencePoints: ProfilePoint[];
-  controlPoints: ProfilePoint[];
-  curvePoints: Vec2[];
-  sections: number;
-  angleStep: number;
-};
+// common function for calculating draw props of compontents building SimpleProfile
+export const calculateDrawProps = (
+  profileRadius: number,
+  maxOffset: number,
+  sections: number
+): DrawProps => {
+  const maxRadius = profileRadius + maxOffset;
+  const strokeWidth = maxRadius / 128.0;
 
-export const distanceSqr = (a: Vec2, b: Vec2): number => {
-  const dx = a.x - b.x;
-  const dy = a.y - b.y;
+  const vbPadding = strokeWidth * 4;
+  const vbMinX = -maxRadius - vbPadding;
+  const vbMinY = -maxRadius - vbPadding;
+  const vbWidth = 2 * (maxRadius + vbPadding);
+  const vbHeight = 2 * (maxRadius + vbPadding);
+  const viewBox = `${vbMinX} ${vbMinY} ${vbWidth} ${vbHeight}`;
 
-  return dx * dx + dy * dy;
-};
+  const sectionAngle = (Math.PI * 2) / sections;
+  const angleStart = -sectionAngle / 2 - Math.PI / 2;
 
-export const distance = (a: Vec2, b: Vec2): number => {
-  return Math.sqrt(distanceSqr(a, b));
-};
-
-export const pointOnCircle = (radius: number, angle: number): Vec2 => {
   return {
-    x: radius * Math.cos(angle),
-    y: radius * Math.sin(angle),
+    strokeWidth: strokeWidth,
+    viewBox: viewBox,
+    sections: sections,
+    angleStart: angleStart,
+    angleStep: sectionAngle,
+    profileRadius: profileRadius,
+    maxOffset: maxRadius,
   };
 };
 
-export const rotate = (
-  point: Vec2,
-  cosAngle: number,
-  sinAngle: number
-): Vec2 => {
-  return {
-    x: point.x * cosAngle - point.y * sinAngle,
-    y: point.x * sinAngle + point.y * cosAngle,
-  };
-};
-
+// generates segment points for a single section
 export const generateSegmentProfilePoints = (
   profileRadius: number,
   minAngle: number,
@@ -87,6 +75,7 @@ export const generateSegmentProfilePoints = (
   return points;
 };
 
+// sorts points by angle and priority (order of source PointSets)
 export const sortProfilePoints = (profilePoints: ProfilePoint[]) => {
   profilePoints.sort((a, b) => {
     if (a.angle < b.angle) {
@@ -107,6 +96,109 @@ export const sortProfilePoints = (profilePoints: ProfilePoint[]) => {
 
     return 0;
   });
+};
+
+// generate curve for a single section
+export const generateProfileSectionCurve = (
+  profileSectionPoints: ProfilePoint[],
+  minSectionAngle: number,
+  maxSectionAngle: number,
+  sizeUnit: SizeUnit
+): Vec2[] => {
+  if (profileSectionPoints.length < 2) {
+    return [];
+  }
+
+  const sectionAngle = maxSectionAngle - minSectionAngle;
+  const sectionAngleCos = Math.cos(sectionAngle);
+  const sectionAngleSin = Math.sin(sectionAngle);
+
+  const sectionFirstPoint = profileSectionPoints[0];
+  const sectionSecondPoint = profileSectionPoints[1];
+  const sectionLastPoint =
+    profileSectionPoints[profileSectionPoints.length - 1];
+
+  const pp: Vec2[] = [
+    rotate(sectionLastPoint.position, sectionAngleCos, -sectionAngleSin), // rotate -sectionAngle (last point in prev section)
+    ...profileSectionPoints.map((point) => point.position),
+    rotate(sectionFirstPoint.position, sectionAngleCos, sectionAngleSin), // rotate sectionAngle (first point in next section)
+    rotate(sectionSecondPoint.position, sectionAngleCos, sectionAngleSin), // rotate sectionAngle (second point in next section)
+  ];
+
+  const minDistance = minDistanceForUnit(sizeUnit);
+  const minDistanceSqr = minDistance * minDistance;
+
+  let prev = pp[1];
+  let prevAngle = Math.atan2(prev.y, prev.x);
+  const points: Vec2[] = [prev];
+
+  for (let i = 1; i < pp.length - 2; i++) {
+    const p0 = pp[i - 1];
+    const p1 = pp[i];
+    const p2 = pp[i + 1];
+    const p3 = pp[i + 2];
+
+    const segmentLength = estimateCatmullRomCurveLength(
+      p0,
+      p1,
+      p2,
+      p3,
+      minDistance
+    );
+
+    const steps = Math.ceil(segmentLength / minDistance);
+    const step = 1 / steps;
+
+    for (let j = 1; j < steps; j++) {
+      const t = (j + 1) * step;
+      let point = catmullRomCurvePoint(p0, p1, p2, p3, t);
+
+      const pointAngle = Math.atan2(point.y, point.x);
+
+      // prev and point are not oriented clockwise, rotate point to fix it
+      if (pointAngle < prevAngle + ANGLE_EPSILON) {
+        const angleDiff = prevAngle + ANGLE_EPSILON - pointAngle;
+        point = rotate(point, Math.cos(angleDiff), Math.sin(angleDiff));
+      }
+
+      // outside of the section
+      if (
+        pointAngle < minSectionAngle - ANGLE_EPSILON ||
+        pointAngle > maxSectionAngle + ANGLE_EPSILON
+      ) {
+        continue;
+      }
+
+      const distSqr = distanceSqr(prev, point);
+
+      // too close to the previous point
+      if (distSqr <= minDistanceSqr) {
+        continue;
+      }
+
+      points.push(point);
+      prev = point;
+      prevAngle = pointAngle;
+    }
+  }
+
+  // no points generated
+  if (points.length == 0) {
+    return [];
+  }
+
+  // last point is too close to the first point == the same point after rotation
+  if (distanceSqr(points[0], points[points.length - 1]) <= minDistanceSqr) {
+    points.pop();
+  }
+
+  // TODO: remove before rollout
+  console.log("generated points: ", points.length);
+
+  const minSimplifyArea = minSimplifyAreaForUnit(sizeUnit);
+  const simplified = simplifyProfilePoints(points, minSimplifyArea);
+
+  return simplified;
 };
 
 export const generateProfile = (
@@ -181,34 +273,5 @@ export const generateProfile = (
     curvePoints: curvePoints,
     sections: sections,
     angleStep: angleStep,
-  };
-};
-
-export const calculateDrawProps = (
-  profileRadius: number,
-  maxOffset: number,
-  sections: number
-): DrawProps => {
-  const maxRadius = profileRadius + maxOffset;
-  const strokeWidth = maxRadius / 128.0;
-
-  const vbPadding = strokeWidth * 4;
-  const vbMinX = -maxRadius - vbPadding;
-  const vbMinY = -maxRadius - vbPadding;
-  const vbWidth = 2 * (maxRadius + vbPadding);
-  const vbHeight = 2 * (maxRadius + vbPadding);
-  const viewBox = `${vbMinX} ${vbMinY} ${vbWidth} ${vbHeight}`;
-
-  const sectionAngle = (Math.PI * 2) / sections;
-  const angleStart = -sectionAngle / 2 - Math.PI / 2;
-
-  return {
-    strokeWidth: strokeWidth,
-    viewBox: viewBox,
-    sections: sections,
-    angleStart: angleStart,
-    angleStep: sectionAngle,
-    profileRadius: profileRadius,
-    maxOffset: maxRadius,
   };
 };
