--- src/lib/types.ts
@@ -1,4 +1,0 @@
-export enum SizeUnit {
-  Centimeter = "cm",
-  Inch = "in",
-}
